Deleted the 'manual' directory and moved the files to the 'particle1d' directory.
