+++ examples/particle1d/particle1d_manual.cc
@@ -1,5 +1,5 @@
 // -----------------------------------------------------------------------------
-#include "drake/examples/particle1d/manual/particle1d_manual.h"
+#include "drake/examples/particle1d/particle1d_manual.h"
 
 // -----------------------------------------------------------------------------
 namespace Manual {
